courses list page

+++ src/main/java/online_learn/constants/PageSizeConst.java
@@ -2,5 +2,5 @@ package online_learn.constants;
 
 public class PageSizeConst {
 
-    public static final long COURSES_PAGE = 6;
+    public static final int COURSES_PAGE = 6;
 }
